adjust includes and imports to use base 'magpie.' package
