+++ magpie/api/api_requests.py
@@ -1,8 +1,8 @@
 from magpie import *
-from definitions.pyramid_definitions import *
-from definitions.ziggurat_definitions import *
-from api.api_except import evaluate_call, verify_param
-from api.api_rest_schemas import *
+from magpie.definitions.pyramid_definitions import *
+from magpie.definitions.ziggurat_definitions import *
+from magpie.api.api_except import evaluate_call, verify_param
+from magpie.api.api_rest_schemas import *
 import models
 
 
@@ -39,7 +39,7 @@ def get_multiformat_delete(request, key, default=None):
 
 def get_permission_multiformat_post_checked(request, service_resource, permission_name_key='permission_name'):
     # import here to avoid circular import error with undefined functions between (api_request, resource_utils)
-    from api.management.resource.resource_utils import check_valid_service_resource_permission
+    from magpie.api.management.resource.resource_utils import check_valid_service_resource_permission
     perm_name = get_value_multiformat_post_checked(request, permission_name_key)
     check_valid_service_resource_permission(perm_name, service_resource, request.db)
     return perm_name
@@ -133,7 +133,7 @@ def get_service_matchdict_checked(request, service_name_key='service_name'):
 
 def get_permission_matchdict_checked(request, service_resource, permission_name_key='permission_name'):
     # import here to avoid circular import error with undefined functions between (api_request, resource_utils)
-    from api.management.resource.resource_utils import check_valid_service_resource_permission
+    from magpie.api.management.resource.resource_utils import check_valid_service_resource_permission
     perm_name = get_value_matchdict_checked(request, permission_name_key)
     check_valid_service_resource_permission(perm_name, service_resource, request.db)
     return perm_name
