+++ magpie/api/management/resource/resource_utils.py
@@ -3,12 +3,12 @@ from common import *
 from models import resource_factory, resource_type_dict, resource_tree_service
 from services import service_type_dict
 from register import sync_services_phoenix
-from definitions.pyramid_definitions import *
-from definitions.ziggurat_definitions import *
-from api.api_rest_schemas import *
-from api.api_requests import *
-from api.api_except import verify_param, evaluate_call, raise_http, valid_http
-from api.management.resource.resource_formats import format_resource
+from magpie.definitions.pyramid_definitions import *
+from magpie.definitions.ziggurat_definitions import *
+from magpie.api.api_rest_schemas import *
+from magpie.api.api_requests import *
+from magpie.api.api_except import verify_param, evaluate_call, raise_http, valid_http
+from magpie.api.management.resource.resource_formats import format_resource
 
 
 def check_valid_service_resource_permission(permission_name, service_resource, db_session):
