+++ magpie/api/management/resource/resource_views.py
@@ -1,9 +1,9 @@
-from api.api_requests import *
-from api.management.service.service_utils import get_services_by_type
-from api.management.service.service_formats import format_service_resources
-from api.management.resource.resource_utils import *
-from api.management.resource.resource_formats import *
-from definitions.pyramid_definitions import view_config
+from magpie.api.api_requests import *
+from magpie.api.management.service.service_utils import get_services_by_type
+from magpie.api.management.service.service_formats import format_service_resources
+from magpie.api.management.resource.resource_utils import *
+from magpie.api.management.resource.resource_formats import *
+from magpie.definitions.pyramid_definitions import view_config
 from common import str2bool
 from register import sync_services_phoenix
 from services import service_type_dict
