+++ magpie/api/management/service/service_utils.py
@@ -1,9 +1,9 @@
 from magpie import models
 from register import SERVICES_PHOENIX_ALLOWED
-from definitions.ziggurat_definitions import *
+from magpie.definitions.ziggurat_definitions import *
 from services import service_type_dict
-from api.api_except import *
-from api.management.group.group_utils import create_group_resource_permission
+from magpie.api.api_except import *
+from magpie.api.management.group.group_utils import create_group_resource_permission
 import os
