+++ magpie/db.py
@@ -1,6 +1,6 @@
 from magpie import MAGPIE_ROOT
-from definitions.alembic_definitions import *
-from definitions.sqlalchemy_definitions import *
+from magpie.definitions.alembic_definitions import *
+from magpie.definitions.sqlalchemy_definitions import *
 from common import print_log
 # noinspection PyCompatibility
 import configparser
