+++ magpie/services.py
@@ -1,8 +1,8 @@
 from magpie import *
 from owsrequest import *
-from definitions.ziggurat_definitions import *
-from definitions.pyramid_definitions import EVERYONE, ALLOW
-from api.api_except import *
+from magpie.definitions.ziggurat_definitions import *
+from magpie.definitions.pyramid_definitions import EVERYONE, ALLOW
+from magpie.api.api_except import *
 import models
