+++ magpie/ui/home/views.py
@@ -1,5 +1,5 @@
-from definitions.pyramid_definitions import *
-from ui.home import add_template_data
+from magpie.definitions.pyramid_definitions import *
+from magpie.ui.home import add_template_data
 
 
 @view_config(route_name='home', renderer='templates/home.mako', permission=NO_PERMISSION_REQUIRED)
