+++ magpie/ui/login/views.py
@@ -1,7 +1,7 @@
 import requests
-from definitions.pyramid_definitions import *
-from ui.management import check_response
-from ui.home import add_template_data
+from magpie.definitions.pyramid_definitions import *
+from magpie.ui.management import check_response
+from magpie.ui.home import add_template_data
 
 
 class ManagementViews(object):
